+++ Electron/vue2/src/TUIRoom/components/Chat/MessageList/index.vue
@@ -20,20 +20,17 @@
 </template>
 
 <script setup lang="ts">
-import { nextTick, onMounted, onUnmounted } from 'vue';
+import { nextTick, onMounted } from 'vue';
 import MessageText from '../MessageTypes/MessageText.vue';
-import TUIRoomEngine, { TUIRoomEvents } from '@tencentcloud/tuiroom-engine-electron';
-import { isMobile }  from '../../../utils/useMediaValue';
+import { isMobile }  from '../../../utils/environment';
 import useMessageList from '../useMessageListHook';
 import { useChatStore } from '../../../stores/chat';
 const chatStore = useChatStore();
 
 const {
   messageAimId,
-  roomEngine,
   messageBottomEl,
   handleMessageListScroll,
-  onReceiveTextMessage,
   messageList,
   getMessageList,
 } = useMessageList();
@@ -55,15 +52,6 @@ onMounted(async () => {
     }
   });
   window.addEventListener('scroll', handleMessageListScroll, true);
-});
-
-TUIRoomEngine.once('ready', () => {
-  roomEngine.instance?.on(TUIRoomEvents.onReceiveTextMessage, onReceiveTextMessage);
-});
-
-onUnmounted(() => {
-  window.removeEventListener('scroll', handleMessageListScroll, true);
-  roomEngine.instance?.off(TUIRoomEvents.onReceiveTextMessage, onReceiveTextMessage);
 });
 </script>