+++ Electron/vue2/src/TUIRoom/components/RoomContent/StreamContainer/index.vue
@@ -156,7 +156,7 @@ watch([() => showStreamList.value.map(item => item.userId), currentPageIndex], (
 });
 
 watch(streamNumber, (val) => {
-  if (currentPageIndex.value > Math.ceil(val / 9) - 1) {
+  if (layout.value === LAYOUT.NINE_EQUAL_POINTS && currentPageIndex.value > Math.ceil(val / 9) - 1) {
     currentPageIndex.value = Math.ceil(val / 9) - 1;
     handleNineEqualPointsLayout();
   }
@@ -529,7 +529,7 @@ const onUserVideoStateChanged = (eventInfo: {
     if (streamType === TUIVideoStreamType.kCameraStream) {
       TUIMessage({
         type: 'warning',
-        message: t('The host has turned off your camera'),
+        message: t('Your camera has been turned off'),
         duration: MESSAGE_DURATION.NORMAL,
       });
       // When the moderator opens the whole staff forbidden to draw,
@@ -577,8 +577,8 @@ const onUserVideoStateChanged = (eventInfo: {
 // 麦位变化
 const onSeatListChanged = (eventInfo:
   { seatList: TUISeatInfo[], seatedList: TUISeatInfo[], leftList: TUISeatInfo[] }) => {
-  const { seatList, seatedList, leftList } = eventInfo;
-  roomStore.updateOnSeatList(seatList, seatedList, leftList);
+  const { seatedList, leftList } = eventInfo;
+  roomStore.updateOnSeatList(seatedList, leftList);
   const leftUserIdList = leftList.map(item => item.userId);
   // 最大屏用户下麦时，需要更换最大屏用户
   if (enlargeStream.value && leftUserIdList.includes(enlargeStream.value?.userId)) {